Moved searchbar related files to it's own folder
